move weird file
