+++ crates/iota-sdk-ffi/src/graphql/api/coins.rs
@@ -9,8 +9,8 @@ use iota_sdk::graphql_client::pagination::PaginationFilter;
 
 use crate::{
     error::Result,
-    graphql::{client::GraphQLClient, pagination::CoinPage},
-    types::{address::Address, graphql::CoinMetadata, struct_tag::StructTag},
+    graphql::{client::GraphQLClient, pagination::CoinPage, query_types::CoinMetadata},
+    types::{address::Address, struct_tag::StructTag},
 };
 
 #[uniffi::export(async_runtime = "tokio")]
