+++ crates/iota-sdk-ffi/src/types/mod.rs
@@ -9,7 +9,6 @@ pub mod digest;
 pub mod events;
 pub mod execution_status;
 pub mod gas;
-pub mod graphql;
 pub mod iota_names;
 pub mod move_package;
 pub mod object;
